# Speed up CheckHighForm