@albert-magyar 2020-05-09
Also, this might help reduce the massive stack footprint I see sometimes in `CheckHighForm` :)